last commit

+++ lib/pages/tabs/documents_tab.dart
@@ -5,6 +5,7 @@ import 'package:bluetooth_app/bloc/bloc.bloc.dart';
 import 'package:bluetooth_app/models/nomenclature.dart';
 import 'package:bluetooth_app/models/product.dart';
 import 'package:bluetooth_app/widgets/text_feild.dart';
+import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 
 class DocumentsTab extends StatefulWidget {
   const DocumentsTab({super.key});
@@ -40,116 +41,7 @@ class _DocumentsTabState extends State<DocumentsTab> {
       body: BlocBuilder<GenericBloc<Nomenclature>, GenericState<Nomenclature>>(
         builder: (context, state) {
           if (state is ItemsLoaded<Nomenclature>) {
-            final archiveCategory = state.items.firstWhere(
-              (nomenclature) => nomenclature.name == 'Архив',
-              orElse: () => Nomenclature(id: '', name: '', isHide: false),
-            );
-
-            final tagCategory = state.items.firstWhere(
-              (nomenclature) => nomenclature.name == 'TAG',
-              orElse: () => Nomenclature(id: '', name: '', isHide: false),
-            );
-
-            otherNomenclatures = state.items
-                .where((nomenclature) =>
-                    nomenclature.name != 'Архив' && nomenclature.name != 'TAG')
-                .toList();
-
-            return CustomScrollView(
-              slivers: [
-                if (archiveCategory.id.isNotEmpty) ...[
-                  const SliverToBoxAdapter(
-                    child: Padding(
-                      padding: EdgeInsets.all(8.0),
-                      child: Text(
-                        'Системные:',
-                        style: TextStyle(
-                            fontSize: 18, fontWeight: FontWeight.bold),
-                      ),
-                    ),
-                  ),
-                  SliverToBoxAdapter(
-                    child: _buildNomenclatureTile(archiveCategory, false,
-                        icon: Icons.archive),
-                  ),
-                  SliverToBoxAdapter(
-                    child: _buildNomenclatureTile(tagCategory, false,
-                        icon: Icons.tag),
-                  ),
-                  const SliverToBoxAdapter(
-                    child: Divider(),
-                  ),
-                ],
-                const SliverToBoxAdapter(
-                  child: Padding(
-                    padding: EdgeInsets.all(8.0),
-                    child: Text(
-                      'Пользовательские:',
-                      style:
-                          TextStyle(fontSize: 18, fontWeight: FontWeight.bold),
-                    ),
-                  ),
-                ),
-                SliverToBoxAdapter(
-                  child: ReorderableListView.builder(
-                    shrinkWrap: true,
-                    onReorder: (int oldIndex, int newIndex) {
-                      setState(() {
-                        if (newIndex > oldIndex) {
-                          newIndex -= 1;
-                        }
-                        bloc.add(ReorderList(newIndex , oldIndex ));
-                      });
-                    },
-                    itemBuilder: (context, index) {
-                      final nomenclature = state.items[index];
-                      return ListTile(
-                        key: ValueKey(nomenclature),
-                        leading: nomenclature.isHide
-                            ? const Icon(Icons.visibility_off, size: 24)
-                            : const Icon(Icons.category),
-                        title: Text("${nomenclature.name} | $index"),
-                        trailing: PopupMenuButton(
-                          onSelected: (value) {
-                            switch (value) {
-                              case 'edit':
-                                _showEditCategoryDialog(nomenclature);
-                                break;
-                              case 'hide':
-                                bloc.add(UpdateItem<Nomenclature>(
-                                  nomenclature.copyWith(
-                                      isHide: !nomenclature.isHide),
-                                ));
-                                break;
-                              case 'delete':
-                                _showDeleteConfirmationDialog(nomenclature);
-                                break;
-                            }
-                          },
-                          itemBuilder: (context) => [
-                            const PopupMenuItem(
-                              value: 'edit',
-                              child: Text('Изменить'),
-                            ),
-                            PopupMenuItem(
-                              value: 'hide',
-                              child: Text(nomenclature.isHide
-                                  ? 'Убрать из скрытых'
-                                  : 'Скрыть'),
-                            ),
-                            const PopupMenuItem(
-                              value: 'delete',
-                              child: Text('Удалить'),
-                            ),
-                          ],
-                        ),
-                      );
-                    },
-                    itemCount: state.items.length,
-                  ),
-                ),
-              ],
-            );
+            return _buildNomenclatureList(state.items);
           }
 
           if (state is ItemOperationFailed<Nomenclature>) {
@@ -223,17 +115,8 @@ class _DocumentsTabState extends State<DocumentsTab> {
               ),
             ),
           ),
-          ReorderableList(
-            onReorder: (int oldIndex, int newIndex) {
-              if (newIndex > oldIndex) {
-                newIndex -= 1;
-              }
-              setState(() {
-                final Nomenclature item = otherNomenclatures.removeAt(oldIndex);
-                otherNomenclatures.insert(newIndex, item);
-              });
-            },
-            itemBuilder: (context, index) {
+          SliverList(
+            delegate: SliverChildBuilderDelegate((context, index) {
               final nomenclature = otherNomenclatures[index];
               return ListTile(
                 key: ValueKey(nomenclature),
@@ -274,8 +157,7 @@ class _DocumentsTabState extends State<DocumentsTab> {
                   ],
                 ),
               );
-            },
-            itemCount: otherNomenclatures.length,
+            }, childCount: otherNomenclatures.length),
           ),
         ],
       );
@@ -289,13 +171,18 @@ class _DocumentsTabState extends State<DocumentsTab> {
         .where((product) => product.category.id == nomenclature.id)
         .toList();
 
+    String product = relatedProducts.length == 1 ? 'продуктом' : 'продуктами';
+    String zeroProduct =
+        'Эта категория не связана ни с одним продуктом. Что Вы хотите сделать?';
     showDialog(
       context: context,
       builder: (context) {
         return AlertDialog(
           title: const Text('Удаление категории'),
           content: Text(
-            'Эта категория связана с ${relatedProducts.length} продуктами. Что вы хотите сделать?',
+            relatedProducts.isNotEmpty
+                ? 'Эта категория связана с ${relatedProducts.length} $product. Что вы хотите сделать?'
+                : zeroProduct,
           ),
           actions: [
             TextButton(
@@ -304,31 +191,46 @@ class _DocumentsTabState extends State<DocumentsTab> {
               },
               child: const Text('Отмена'),
             ),
-            TextButton(
-              onPressed: () {
-                for (var product in relatedProducts) {
-                  productsBox.delete(product.id);
-                }
-                bloc.add(DeleteItem<Nomenclature>(nomenclature.id));
-                Navigator.pop(context);
-              },
-              child: const Text('Удалить вместе с категорией'),
-            ),
-            TextButton(
-              onPressed: () async {
-                var archiveCategory =
-                    Hive.box<Nomenclature>('nomenclature_box').get('archive');
-                for (var product in relatedProducts) {
-                  var updatedProduct =
-                      product.copyWith(category: archiveCategory);
-                  await productsBox.put(product.id, updatedProduct);
-                }
+            if (relatedProducts.isNotEmpty) ...[
+              TextButton(
+                onPressed: () {
+                  for (var product in relatedProducts) {
+                    productsBox.delete(product.id);
+                  }
+                  bloc.add(DeleteItem<Nomenclature>(nomenclature.id));
+                  Navigator.pop(context);
+                },
+                child: const Text('Удалить вместе с категорией'),
+              ),
+              TextButton(
+                onPressed: () async {
+                  var archiveCategory =
+                      Hive.box<Nomenclature>('nomenclature_box').get('archive');
+                  for (var product in relatedProducts) {
+                    var updatedProduct =
+                        product.copyWith(category: archiveCategory);
+                    await productsBox.put(product.id, updatedProduct);
+                  }
 
-                bloc.add(DeleteItem<Nomenclature>(nomenclature.id));
-                Navigator.pop(context);
-              },
-              child: const Text('Архивировать продукты и удалить категорию'),
-            ),
+                  bloc.add(DeleteItem<Nomenclature>(nomenclature.id));
+
+                  if (context.mounted) {
+                    Navigator.pop(context);
+                  }
+                },
+                child: const Text('Архивировать продукты и удалить категорию'),
+              ),
+            ] else
+              TextButton(
+                onPressed: () {
+                  for (var product in relatedProducts) {
+                    productsBox.delete(product.id);
+                  }
+                  bloc.add(DeleteItem<Nomenclature>(nomenclature.id));
+                  Navigator.pop(context);
+                },
+                child: const Text('Удалить'),
+              ),
           ],
         );
       },
@@ -387,58 +289,55 @@ class _DocumentsTabState extends State<DocumentsTab> {
   }
 
   void _showAddCategoryModal(BuildContext context) {
-    showModalBottomSheet(
+    showBarModalBottomSheet(
+      backgroundColor: Colors.white,
       context: context,
       builder: (context) {
-        return BottomSheet(
-          onClosing: () {},
-          builder: (context) {
-            return Scaffold(
-              body: Padding(
-                padding: const EdgeInsets.all(8.0),
-                child: CustomScrollView(
-                  slivers: [
-                    const SliverAppBar(
-                      title: Text('Добавление категории'),
-                      centerTitle: true,
-                      automaticallyImplyLeading: false,
-                    ),
-                    SliverToBoxAdapter(
-                      child: TextInput(
-                        controller: nameController,
-                        hintText: 'Специи',
-                        labelText: 'Название',
-                      ),
-                    ),
-                    const SliverToBoxAdapter(
-                      child: SizedBox(
-                        height: 30,
-                      ),
-                    ),
-                    SliverToBoxAdapter(
-                      child: ElevatedButton(
-                        onPressed: () {
-                          if (nameController.text.isEmpty) {
-                            setState(() {
-                              valid = true;
-                            });
-                            return;
-                          }
-                          bloc.add(AddItem<Nomenclature>(
-                            Nomenclature(
-                                name: nameController.text, isHide: false),
-                          ));
-                          nameController.clear();
-                          Navigator.pop(context);
-                        },
-                        child: const Text('Добавить'),
-                      ),
-                    ),
-                  ],
+        return SizedBox(
+          height: 500,
+          child: Padding(
+            padding: const EdgeInsets.all(8.0),
+            child: CustomScrollView(
+              slivers: [
+                const SliverAppBar(
+                  title: Text('Добавление категории'),
+                  centerTitle: true,
+                  backgroundColor: Colors.transparent,
+                  automaticallyImplyLeading: false,
                 ),
-              ),
-            );
-          },
+                SliverToBoxAdapter(
+                  child: TextInput(
+                    controller: nameController,
+                    hintText: 'Специи',
+                    labelText: 'Название',
+                  ),
+                ),
+                const SliverToBoxAdapter(
+                  child: SizedBox(
+                    height: 30,
+                  ),
+                ),
+                SliverToBoxAdapter(
+                  child: ElevatedButton(
+                    onPressed: () {
+                      if (nameController.text.isEmpty) {
+                        setState(() {
+                          valid = true;
+                        });
+                        return;
+                      }
+                      bloc.add(AddItem<Nomenclature>(
+                        Nomenclature(name: nameController.text, isHide: false),
+                      ));
+                      nameController.clear();
+                      Navigator.pop(context);
+                    },
+                    child: const Text('Добавить'),
+                  ),
+                ),
+              ],
+            ),
+          ),
         );
       },
     );
@@ -464,7 +363,8 @@ class _DocumentsTabState extends State<DocumentsTab> {
                 ElevatedButton(
                   onPressed: () {
                     bloc.add(UpdateItem<Nomenclature>(
-                      nomenclature.copyWith(id: nomenclature.id, name: editController.text),
+                      nomenclature.copyWith(
+                          id: nomenclature.id, name: editController.text),
                     ));
                     editController.clear();
                     Navigator.pop(context);
